Windows port: experimental deepseek_prefill: fix narrowing conversions
- attn_res_gather_softmax: static_cast CoreCoord x/y (size_t) to
  uint32_t in SetRuntimeArgs vector
- combine, dispatch: static_cast packet_header_size_bytes (size_t) to
  uint32_t in CBFormatDescriptor::page_size designated initializer;
  static_cast MeshCoordinateRange/MeshShape accessors (size_t) to
  uint32_t in runtime-args initializer lists
- zero_padded_kv_cache: static_cast CB index (uint32_t) to uint8_t in
  CBFormatDescriptor::buffer_index designated initializer

Note: attn_res_gather_softmax, attn_res_weighted_reduce_nc,
inbound_socket_service_sync, outbound_socket_service_sync,
pack_scaled_fp8_kv_cache, per_token_cast_to_fp8,
rotary_embedding_indexed, update_padded_kv_cache, and
zero_padded_kv_cache still fail to build in this tree: their
generated .obj/.lib output paths exceed the 260-char MAX_PATH limit
under .build/windows-msvc, causing C1083/LNK1104 regardless of source
correctness. This needs a build-tree-location fix, not a source fix.

+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/zero_padded_kv_cache/device/zero_padded_kv_cache_device_operation.cpp
@@ -362,7 +362,9 @@ tt::tt_metal::ProgramDescriptor ZeroPaddedKvCacheDeviceOperation::ProgramFactory
         desc.cbs.push_back(CBDescriptor{
             .total_size = npages * page,
             .core_ranges = all_cores,
-            .format_descriptors = {{CBFormatDescriptor{.buffer_index = index, .data_format = fmt, .page_size = page}}},
+            .format_descriptors = {
+                {CBFormatDescriptor{
+                    .buffer_index = static_cast<uint8_t>(index), .data_format = fmt, .page_size = page}}},
         });
     };
     add_cb(kSrcCbIndex, cache_format, cache_tile_size, Wt);
